refs #7310. Fix bug and add additional tests.
Tests added which will fail in the PeakTransform names get out of sync with the options on the PeaksInRegion algorithm.

+++ Code/Mantid/MantidQt/SliceViewer/src/PeakOverlayMultiSphere.cpp
@@ -133,7 +133,7 @@ namespace MantidQt
 
     void PeakOverlayMultiSphere::movePosition(PeakTransform_sptr transform)
     {
-      for(int i = 0; i < m_physicalPeaks.size(); ++i)
+      for(size_t i = 0; i < m_physicalPeaks.size(); ++i)
       { 
         m_physicalPeaks[i]->movePosition(transform);
       }
